simplify jest test, thanks to new proxying
