+++ jest.config.js
@@ -3,7 +3,4 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   collectCoverage: false,
-  transform: {
-    'src/snarky.js': 'ts-jest',
-  },
 };
